results/requirements.txt
numpy==1.25.1
pandas==2.0.3
PyYAML==6.0.1
tabulate==0.9.0
